feat(backend): move AFS processing logic to async consumer
