+++ backend/core/src/application-flagged-sets/application-flagged-sets.module.ts
@@ -11,10 +11,11 @@ import { BullModule } from "@nestjs/bull"
 import { AFSProcessingQueueNames } from "./constants/applications-flagged-sets-constants"
 import { ConfigService } from "@nestjs/config"
 import { SharedModule } from "../shared/shared.module"
+import { ListingRepository } from "../listings/repositories/listing.repository"
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([ApplicationFlaggedSet, Application]),
+    TypeOrmModule.forFeature([ApplicationFlaggedSet, Application, ListingRepository]),
     AuthModule,
     BullModule.forRootAsync({
       imports: [SharedModule],
