+++ backend/core/src/application-flagged-sets/application-flagged-sets.service.ts
@@ -23,6 +23,9 @@ import { Rule } from "./types/rule-enum"
 import { ApplicationFlaggedSetResolveDto } from "./dto/application-flagged-set-resolve.dto"
 import { PaginatedApplicationFlaggedSetQueryParams } from "./paginated-application-flagged-set-query-params"
 import { ListingStatus } from "../listings/types/listing-status-enum"
+import { InjectQueue } from "@nestjs/bull"
+import { AFSProcessingQueueNames } from "./constants/applications-flagged-sets-constants"
+import { Queue } from "bull"
 
 @Injectable({ scope: Scope.REQUEST })
 export class ApplicationFlaggedSetsService {
@@ -32,7 +35,8 @@ export class ApplicationFlaggedSetsService {
     @InjectRepository(Application)
     private readonly applicationsRepository: Repository<Application>,
     @InjectRepository(ApplicationFlaggedSet)
-    private readonly afsRepository: Repository<ApplicationFlaggedSet>
+    private readonly afsRepository: Repository<ApplicationFlaggedSet>,
+    @InjectQueue(AFSProcessingQueueNames.afsProcessing) private afsProcessingQueue: Queue,
   ) {}
   async listPaginated(queryParams: PaginatedApplicationFlaggedSetQueryParams) {
     const results = await paginate<ApplicationFlaggedSet>(
@@ -114,254 +118,7 @@ export class ApplicationFlaggedSetsService {
     })
   }
 
-  async onApplicationSave(newApplication: Application, transactionalEntityManager: EntityManager) {
-    for (const rule of [Rule.email, Rule.nameAndDOB]) {
-      await this.updateApplicationFlaggedSetsForRule(
-        transactionalEntityManager,
-        newApplication,
-        rule
-      )
-    }
-  }
-
-  private async _getAfsesContainingApplicationId(
-    queryRunnery: QueryRunner,
-    applicationId: string
-  ): Promise<Array<{ application_flagged_set_id: string }>> {
-    const metadataArgsStorage = getMetadataArgsStorage().findJoinTable(
-      ApplicationFlaggedSet,
-      "applications"
-    )
-    const applicationsJunctionTableName = metadataArgsStorage.name
-    const query = `
-      SELECT DISTINCT application_flagged_set_id FROM ${applicationsJunctionTableName}
-      WHERE applications_id = $1
-  `
-    return await queryRunnery.query(query, [applicationId])
-  }
-
-  async onApplicationUpdate(
-    newApplication: Application,
-    transactionalEntityManager: EntityManager
-  ) {
-    const transApplicationsRepository = transactionalEntityManager.getRepository(Application)
-    newApplication.markedAsDuplicate = false
-    await transApplicationsRepository.save(newApplication)
-
-    const transAfsRepository = transactionalEntityManager.getRepository(ApplicationFlaggedSet)
-
-    const afsIds = await this._getAfsesContainingApplicationId(
-      transAfsRepository.queryRunner,
-      newApplication.id
-    )
-    const afses = await transAfsRepository.find({
-      where: { id: In(afsIds.map((afs) => afs.application_flagged_set_id)) },
-      relations: ["applications"],
-    })
-    const afsesToBeSaved: Array<ApplicationFlaggedSet> = []
-    const afsesToBeRemoved: Array<ApplicationFlaggedSet> = []
-    for (const afs of afses) {
-      afs.status = FlaggedSetStatus.flagged
-      afs.resolvedTime = null
-      afs.resolvingUser = null
-      const applicationIndex = afs.applications.findIndex(
-        (application) => application.id === newApplication.id
-      )
-      afs.applications.splice(applicationIndex, 1)
-      if (afs.applications.length > 1) {
-        afsesToBeSaved.push(afs)
-      } else {
-        afsesToBeRemoved.push(afs)
-      }
-    }
-    await transAfsRepository.save(afsesToBeSaved)
-    await transAfsRepository.remove(afsesToBeRemoved)
-
-    await this.onApplicationSave(newApplication, transactionalEntityManager)
-  }
-
-  async fetchDuplicatesMatchingRule(
-    transactionalEntityManager: EntityManager,
-    application: Application,
-    rule: Rule
-  ) {
-    switch (rule) {
-      case Rule.nameAndDOB:
-        return await this.fetchDuplicatesMatchingNameAndDOBRule(
-          transactionalEntityManager,
-          application
-        )
-      case Rule.email:
-        return await this.fetchDuplicatesMatchingEmailRule(transactionalEntityManager, application)
-    }
-  }
-
-  async updateApplicationFlaggedSetsForRule(
-    transactionalEntityManager: EntityManager,
-    newApplication: Application,
-    rule: Rule
-  ) {
-    const applicationsMatchingRule = await this.fetchDuplicatesMatchingRule(
-      transactionalEntityManager,
-      newApplication,
-      rule
-    )
-    const transAfsRepository = transactionalEntityManager.getRepository(ApplicationFlaggedSet)
-    const visitedAfses = []
-    const afses = await transAfsRepository
-      .createQueryBuilder("afs")
-      .leftJoin("afs.applications", "applications")
-      .select(["afs", "applications.id"])
-      .where(`afs.listing_id = :listingId`, { listingId: newApplication.listing.id })
-      .andWhere(`rule = :rule`, { rule })
-      .getMany()
-
-    for (const matchedApplication of applicationsMatchingRule) {
-      const afsesMatchingRule = afses.filter((afs) =>
-        afs.applications.map((app) => app.id).includes(matchedApplication.id)
-      )
-
-      if (afsesMatchingRule.length === 0) {
-        const newAfs: DeepPartial<ApplicationFlaggedSet> = {
-          rule: rule,
-          resolvedTime: null,
-          resolvingUser: null,
-          status: FlaggedSetStatus.flagged,
-          applications: [newApplication, matchedApplication],
-          listing: newApplication.listing,
-        }
-        await transAfsRepository.save(newAfs)
-      } else if (afsesMatchingRule.length === 1) {
-        for (const afs of afsesMatchingRule) {
-          if (visitedAfses.includes(afs.id)) {
-            return
-          }
-          visitedAfses.push(afs.id)
-          afs.applications.push(newApplication)
-          await transAfsRepository.save(afs)
-        }
-      } else {
-        console.error(
-          "There should be up to one AFS matching a rule for given application, " +
-            "probably a logic error when creating AFSes"
-        )
-      }
-    }
-  }
-
-  private async fetchDuplicatesMatchingEmailRule(
-    transactionalEntityManager: EntityManager,
-    newApplication: Application
-  ) {
-    const transApplicationsRepository = transactionalEntityManager.getRepository(Application)
-    return await transApplicationsRepository.find({
-      select: ["id"],
-      where: (qb: SelectQueryBuilder<Application>) => {
-        qb.where("Application.id != :id", {
-          id: newApplication.id,
-        })
-          .andWhere("Application.listing.id = :listingId", {
-            listingId: newApplication.listing.id,
-          })
-          .andWhere("Application__applicant.emailAddress = :emailAddress", {
-            emailAddress: newApplication.applicant.emailAddress,
-          })
-          .andWhere("Application.status = :status", { status: "submitted" })
-      },
-    })
-  }
-
-  private async fetchDuplicatesMatchingNameAndDOBRule(
-    transactionalEntityManager: EntityManager,
-    newApplication: Application
-  ) {
-    const transApplicationsRepository = transactionalEntityManager.getRepository(Application)
-    const firstNames = [
-      newApplication.applicant.firstName,
-      ...newApplication.householdMembers.map((householdMember) => householdMember.firstName),
-    ]
-
-    const lastNames = [
-      newApplication.applicant.lastName,
-      ...newApplication.householdMembers.map((householdMember) => householdMember.lastName),
-    ]
-
-    const birthMonths = [
-      newApplication.applicant.birthMonth,
-      ...newApplication.householdMembers.map((householdMember) => householdMember.birthMonth),
-    ]
-
-    const birthDays = [
-      newApplication.applicant.birthDay,
-      ...newApplication.householdMembers.map((householdMember) => householdMember.birthDay),
-    ]
-
-    const birthYears = [
-      newApplication.applicant.birthYear,
-      ...newApplication.householdMembers.map((householdMember) => householdMember.birthYear),
-    ]
-
-    return await transApplicationsRepository.find({
-      select: ["id"],
-      where: (qb: SelectQueryBuilder<Application>) => {
-        qb.where("Application.id != :id", {
-          id: newApplication.id,
-        })
-          .andWhere("Application.listing.id = :listingId", {
-            listingId: newApplication.listing.id,
-          })
-          .andWhere("Application.status = :status", { status: "submitted" })
-          .andWhere(
-            new Brackets((subQb) => {
-              subQb.where("Application__householdMembers.firstName IN (:...firstNames)", {
-                firstNames: firstNames,
-              })
-              subQb.orWhere("Application__applicant.firstName IN (:...firstNames)", {
-                firstNames: firstNames,
-              })
-            })
-          )
-          .andWhere(
-            new Brackets((subQb) => {
-              subQb.where("Application__householdMembers.lastName IN (:...lastNames)", {
-                lastNames: lastNames,
-              })
-              subQb.orWhere("Application__applicant.lastName IN (:...lastNames)", {
-                lastNames: lastNames,
-              })
-            })
-          )
-          .andWhere(
-            new Brackets((subQb) => {
-              subQb.where("Application__householdMembers.birthMonth IN (:...birthMonths)", {
-                birthMonths: birthMonths,
-              })
-              subQb.orWhere("Application__applicant.birthMonth IN (:...birthMonths)", {
-                birthMonths: birthMonths,
-              })
-            })
-          )
-          .andWhere(
-            new Brackets((subQb) => {
-              subQb.where("Application__householdMembers.birthDay IN (:...birthDays)", {
-                birthDays: birthDays,
-              })
-              subQb.orWhere("Application__applicant.birthDay IN (:...birthDays)", {
-                birthDays: birthDays,
-              })
-            })
-          )
-          .andWhere(
-            new Brackets((subQb) => {
-              subQb.where("Application__householdMembers.birthYear IN (:...birthYears)", {
-                birthYears: birthYears,
-              })
-              subQb.orWhere("Application__applicant.birthYear IN (:...birthYears)", {
-                birthYears: birthYears,
-              })
-            })
-          )
-      },
-    })
+  public async scheduleAfsProcessing() {
+    return this.afsProcessingQueue.add(null, {})
   }
 }
