+++ backend/core/src/listings/listings.service.ts
@@ -1,11 +1,4 @@
-import {
-  HttpException,
-  HttpStatus,
-  Inject,
-  Injectable,
-  NotFoundException,
-  Scope,
-} from "@nestjs/common"
+import { HttpException, HttpStatus, Inject, Injectable, NotFoundException, Scope } from "@nestjs/common"
 import { InjectRepository } from "@nestjs/typeorm"
 import { Pagination } from "nestjs-typeorm-paginate"
 import { In, Repository } from "typeorm"
@@ -31,6 +24,7 @@ import { AuthzService } from "../auth/services/authz.service"
 import { Request as ExpressRequest } from "express"
 import { REQUEST } from "@nestjs/core"
 import { User } from "../auth/entities/user.entity"
+import { ApplicationFlaggedSetsService } from "../application-flagged-sets/application-flagged-sets.service"
 
 type OrderByConditionData = {
   orderBy: string
@@ -45,7 +39,8 @@ export class ListingsService {
     @InjectRepository(AmiChart) private readonly amiChartsRepository: Repository<AmiChart>,
     private readonly translationService: TranslationsService,
     private readonly authzService: AuthzService,
-    @Inject(REQUEST) private req: ExpressRequest
+    @Inject(REQUEST) private req: ExpressRequest,
+    private readonly afsService: ApplicationFlaggedSetsService,
   ) {}
 
   private getFullyJoinedQueryBuilder() {
@@ -191,6 +186,10 @@ export class ListingsService {
       }
     })
     listingDto.unitsAvailable = availableUnits
+
+    if (listing.status == ListingStatus.active && listingDto.status == ListingStatus.closed) {
+      await this.afsService.scheduleAfsProcessing()
+    }
 
     Object.assign(listing, {
       ...listingDto,
